+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Complete list:
 - `ConfigPath` class is now split in `ConfigCatalog` and `ConfigContext` classes (#2926)
 - Fix no kind case for aqua analysis (#3032)
 - Implement new DVC remote from ECMWF (#3031)
+- Expand use of `DaskCluster()` class to `Drop()` to control dask (#3037)
 
 ## [v1.0.1]